+++ changelog.md
@@ -3,3 +3,4 @@
 ## Next
 
 * Initial release
+* Show native promise tests as skipped where no native promise implementation (node v0.10)